perf: perf excel comp code
