+++ src/components/Excel/index.ts
@@ -1,4 +1,4 @@
-export { default as ImportExcel } from './src/ImportExcel';
+export { default as ImportExcel } from './src/ImportExcel.vue';
 export { default as ExportExcelModel } from './src/ExportExcelModel.vue';
 
 export { jsonToSheetXlsx, aoaToSheetXlsx } from './src/Export2Excel';
